Compilation fix
git-svn-id: file:///home/roberto/Git/svn@10391 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: a0abfdbb2c716109ab6246e790fe082892864aff

+++ trunk/xmipp/libraries/data/metadata.h
@@ -330,7 +330,7 @@ public:
      */
     bool setPrecission(int _precision)
     {
-        precision = pow (10,_precision);
+        precision = (int)pow (10,_precision);
     }
 
     /** Set to false for row format (parameter files).
