+++ trunk/xmipp/libraries/data/numerical_tools.h
@@ -126,7 +126,7 @@ public:
         if (x>xmax) return 0;
         else
         {
-            int iaux=round(x*ixstep);
+            int iaux=(int)round(x*ixstep);
             return DIRECT_A1D_ELEM(v,iaux);
         }
     }
